fixed delete route

+++ pastebin/routes.py
@@ -62,6 +62,9 @@ def paste_raw_view(paste_hash):
 def paste_delete(paste_hash):
     paste = utils.get_paste_from_hash(paste_hash)
 
+    if not utils.check_if_paste_exists(paste):
+        return redirect(url_for("pastebin.not_found"))
+
     if not utils.delete_paste_if_user_is_author(paste):
         return redirect(url_for("pastebin.paste_view", paste_hash=paste.paste_hash))
